aword

+++ xmega/unit_testing/i2c_master/src/i2c.h
@@ -7,7 +7,7 @@ status_code_t i2c_send_double(TWI_t *twi, uint8_t addr, uint8_t memory, float co
 status_code_t i2c_send_array(TWI_t *twi, uint8_t addr, uint8_t memory, int arraysize, void *array);
 uint8_t i2c_read(TWI_t *twi, uint8_t addr, uint8_t memory);
 uint16_t i2c_read_word(TWI_t *twi, uint8_t addr, uint8_t memory);
-uint32_t i2c_read_double(TWI_t *twi, uint8_t addr, uint8_t memory);
+float i2c_read_double(TWI_t *twi, uint8_t addr, uint8_t memory);
 status_code_t i2c_read_array(TWI_t *twi, uint8_t addr, uint8_t memory, int arraysize, void *array);
 
 #endif /* I2C_H_ */
